+++ font/Tables/CMap/CMapFormat8.cs
@@ -32,7 +32,7 @@ public class CMapFormat8 : ICMapFormat, ICharToGID
             Language = language,
             Is32 = is32,
             NumberOfGroups = num_of_groups,
-            Groups = [.. Lists.Repeat(() => (stream.ReadUIntByBigEndian(), stream.ReadUIntByBigEndian(), stream.ReadUIntByBigEndian())).Take((int)num_of_groups)],
+            Groups = [.. Lists.Repeat(() => (stream.ReadUIntByBigEndian(), stream.ReadUIntByBigEndian(), stream.ReadUIntByBigEndian())).Take(num_of_groups)],
         };
     }
